super roast bot/requirements.txt
streamlit
groq
python-dotenv
faiss-cpu
sentence-transformers
langchain
tiktoken
numpy